save upper bound of 95% confidence interval
A (one-sided) permutation test p-value is simply the number of test statistic values, each corresponding to a permutation sample, at least as large as that based on the observed data samples; therefore, the estimated pvalue based on conventional Monte Carlo sampling is simply an estimated proportion distributed binomially. The normal approximation to the binomial distribution allows one easily to obtain specified levels of precision for this estimate, based on the standard error (Opdyke, 2003).